+++ src/Button/Button.stories.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { withKnobs, text, select, boolean } from '@storybook/addon-knobs'
 import Button from './index'
-import LinkButton from '../LinkButton'
+import LinkButton from './LinkButton'
 
 import { STORY_PARAMS } from '../globals'
 import theme from '../theme'
